outsource init code for IntervalArithmetic
